Updated SQL with join table and sample data
Fixed features and items for many to many relationship

Added sample data to be used in testing the application and for tutoring
purposes.

Todo:

- implement DAO style access for the various items and such. Create code
  that generates the form for lost/found based on the feature types and
  features of a possible item

- create a search utility that allows for searching based on the
  features and etc.

it will probably be good to implement a view layer or something to
construct the form pieces for a feature type based on the is_multi and
is_dropdown fields.