Removed unneccesary form from featuretypeform
Also moved models to shared directory since it makes more sense for it
to be there.

Original intent of featuretypeform was to make it possible to have the
UI dynamically constructed by the feature types and entities given,
kinda like a weird window manager in some ways. I'm going to focus on
functionality first and then try to do this.