Added Event list and retrieval method
getEvents can take nothing -> returns all events
or it can take true or false for whether or not to return just enabled
or disabled events.

Updated models to not have constructors because I forgot for a second
that pdo sets the fields before calling the constructor when it performs
the row mapping onto the class.