Updated event update form, added delete func
Added function to delete by list of id's into EventService, just doing
it one by one right now because performance is not my concern right now
and the number of times events will be deleted is likely to by low.

Changed event update form so that the information comes through as
pretty usable, Basically like this for the POST data:

Array
(
    [events] => Array
        (
            [1] => Array
                (
                    [id] => 1
                    [name] => Sample Event
                    [enabled] => true
                )

            [2] => Array
                (
                    [id] => 2
                    [name] => Test
                )

        )

    [delete] => Array
        (
            [0] => 2
        )

)

events => array of events, enabled only shows if the event is enabled
just like the regular form.
delete => array of ids to be passed to the event service function we
created.

Also updated url config