Skeleton for update-feature-types
Added conditional error message

Created skeleton of idea for update feature types. Added url config for
the updating of feature types, removed the unneccesary code to add
beloning features to the feature types as I've decided to not use a
nested UI like that.

POST data coming in will look something like this, (along with delete
id's if applicable)

Array
(
    [featureTypes] => Array
        (
            [1] => Array
                (
                    [id] => 1
                    [name] => Color
                    [is_] => multi
                )

            [2] => Array
                (
                    [id] => 1
                    [name] => Lost in
                    [is_] => dropdown
                )

            [3] => Array
                (
                    [id] => 1
                    [name] => Lost on
                    [is_] => dropdown
                )

            [4] => Array
                (
                    [id] => 1
                    [name] => Item Type
                    [is_] => dropdown
                )

        )

    [delete] => Array
        (
            [0] => 4
        )

)