Working on update features. still needs work.
added todo for the rest of it, but skeleton is in place.

I think I'm going to want to refactor this afterwards mainly because all
the updates actions function on the same principle, so I should be able
to make a generic controller or something to use in the same way that
the services could probably be generified.