Able to create lost items.
Ability to create lost items is done in the lost form. I should really
just move it to be a report item form with the user choosing whether it
is lost or not. That would make it simpler to do, though the language
does need to be set appropriately in that case.

- new service in core (Item)
- new action (report item)
- new functional page (lost)