Began work on Inventory Page.
Inventory page will be used to view and mark items as found or lost. I
suppose it could also have a small amount of rudimentary javascript
searched too which could simply search the description/titles without
the additional overhead of the actual search page.

Though since it will be similar, I may just make a function of some kind
or template that takes the list of items and displays them with links to
individual items. Then we can feed a list of items whether they be from
the regular inventory page or the search page into it and enjoy.