Added search box and links to adv search.
Inventory page now has a nice search box and some quick links. Need to
make the text search functional and a results page, then need to make a
search page and probably after that a standalone item page to link to
for all these items so you can see the details.